+++ contracts/vm/LibStackTop.sol
@@ -105,7 +105,11 @@ library LibStackTop {
         }
     }
 
-    function up(StackTop stackTop_, uint n_) internal pure returns (StackTop) {
+    function up(StackTop stackTop_, uint256 n_)
+        internal
+        pure
+        returns (StackTop)
+    {
         unchecked {
             return StackTop.wrap(StackTop.unwrap(stackTop_) + 0x20 * n_);
         }
